Added "flipColors" parameter to the AddTexture() method that takes memory images (read by other than SOIL), as part of adding support for OpenCV.
OpenCV uses BGR.  Yes, that's lame, but it's historical, for the same reason that the space shuttle's external fuel tanks are the width of 2 horses asses.  

Thankfully OpenGL can cope, but it has to be told to use a different format when loading texture images.

So I added "flipColors", and it's defaulted to false.

The XGLTexQuad() constructor need the argument too, and it's also defaulted to false.

Added the OpenCVTest project to experiment with while getting this working.