XGLShape is now derived from XGLBuffer, instead of having an XGLBuffer as a child member variable.
This was done as part of the project to eliminate the XGL Singleton.  Turns out not everything I learned at Roku was good practice!  Most folks agree Singletons are evil, something I agreed with even when I chose to use the pattern.  It was an easy work around whose time has come to be eliminated.