In the process of switching XGLLight over to full uniform buffer object implementation.  Work in progress, but this one actually works without too much changing of CPU <-> shader interface code.
The vertex color attribute is still vec3.

The uniform buffer object with std140 layout in the shader pads the vec3 to vec4, but CPU code (XGL) is still using vec3 for position and color.  I padded XGLLight with extra floats to get things to line with the shader layout for this commit.

I'll change back to vec4 for color and position vertex attributes eventually.  Probably Real Soon Now (tm)