Connected OpenAL to AVPlayer AudioStreamThread.
Changed back to GoProVideo.

Using XAVBuffer in VideoStreamThread now results is VERY smooth video playback.

However, CPU usage went way up in this version.  Not sure if it's the video resolution, or all of the changes I made.  I want to save this because it works great, but I'm going to immediately revert to the previous commit and try and figure out what happened to cause the huge jump in CPU usage.