XGLFramebuffer to Filemapping shared memory object (Windows specific) works!!
Well first light anyway.  The usual rectangle shenanigans must be dealt with.

Other DirectShow applications tell the source filter what resolution they want, rather than being forced to cope with what the source filter wants to send.  Didn't know that but not surprised.

Shouldn't be to much of a challenge to deal with.