Significant refactoring to   a) take the model matrix out of the ShaderMatrixData UBO, and   b) get XGLMaterial working per-shape.
Material attribute uniform locations are fetched after a shader is compiled, and stored with the shader.

When a shape is created, it gets a copy of the material attribute locations from the shader it's associated with. This makes XGLMaterial::Bind() easier to code, since it knows now where the uniforms are within each shader.

XGLMaterial::Bind() actually works now.