Finally!!!
Broke the code on how to render to multiple GL_COLORATTACHMENT buffers, and then how to utilize those results within a single shader.

This empowers video signal processing per pixel in real time, all on the GPU.

Specifically, it allows for comparing the most recent incoming video frame with the previous video frame, which allows for cool computer vision techniques that I've been envisioning.

Been after this for awhile. Super stoked that it's finally here!