Cracked the window-relative coordinates nut. Just make a glm::vec4 from the normalized screen coordinates, and multiply that by an inverse of the model matrix.
Found this on StackOverflow http://stackoverflow.com/questions/9863875/glm-how-to-transpose-a-vector:
"So if you have a mat4 and do mat4*vec4, your vec4 is considered a column vector. If you do vec4*mat4, it is considered a row vector."

I find the documentation for GLM less than helpful, because I'm not a guru when it comes to the math involved.  I get the higher level concepts though and that seems to be enough to get by.