AntTweakBar integration is now functional, at least visually. '~' key toggles as with XGLGuiCanvas. Mouse events (motion & buttons) are connected, keyboard isn't yet. ATBShape constructor now takes a pointer to ExampleXGL, and thus can do all the calls to connect itself to the top-level XGL GuiXXX functions.
Added a callback mechanism to XGLProjector so that ATBShape::reshape() will be called when the window resizes.  This needs to be enhanced to a std::vector<> of callbacks, so that multiple reshape() callbacks can be added.  Don't know yet of an immediately pending need for such, but might as well do it now, before I lose my thoughts about it.