Corrected buffer sizing bug in XAVSrc and VideoStreamThread. The memcpy() calls in VideoStreamThread were based on the size of the destination array, NOT the size of the source buffer.  Thus if the destination size was too big, an access violation occurred during those calls.
My objective was to be able to specify a constant size of the destination buffers that is likely to NEVER be exceeded, and only use the portion necessary for the size of the video actually being decoded.  That is now working correctly.

I should add some specific exception handling in case someone violates the max size.