Refactored the orthographic projection stuff for the GUI...
Added orthoProjection to XGLShaderMatrixData, added XGLProjector::GetOrthoMatrix(), calling it from XGL::RenderScene() to set it in the Matrix UBO, modified the MatrixData UBO in "ortho" and "ortho-tex" shaders, removed orthoProjection from XGLGuiCanvas.

So now, all that is needed is to specify a shader that makes use of orthoProjection in the vertex shader as "ortho" & "ortho-tex" do and XGLGuiCanvas shapes will be displayed correctly without any shenanigans.