Adding c_library_v1_unsorted_big_endian...
After much investigation, it turns out that the pymavlink header generation system doesn't appear generating C header files that are correct for *my* ArduPilotMega.  

Mine appears to be responding to v1.0 commands but with the fields in the payload portion of the command unsorted, ie: in they're in the order specified in the mavlink source XML files.  

This was determined by using a USB sniffer and looking at what MissionPlanner emits to the APM to get it to respond. From there I determined that the call to mavlink_msg_request_data_stream_encode() was emitting things in the wrong order.  I traced it back to the pymavlink/generator/mavparse.py source file, which sets various configuration variables according to the specified protocol version.  I modified that file to do what I think is needed, and the result is this C header set.  Specifically, I set self.sort_fields and self.little_endian to false for PROTOCOL_1_0 on lines 192 and 193 respectively, then ran "python -m pymavgen.tools.mavgen" specifying version 1.0.