Success!
Visualizing the orientation of the APM flight controller by parsing the ATTITUDE MAVLINK message and turning it into the appropriate rotation matrix for an XGLShape object.

There's a problem at shutdown time: the XMavlink::Listener for the ATTITUDE message doesn't cope with the deletion of the XGLShape it's modifying, and will crash if the timing doesn't work out in our favor.  

Needs to be fixed, after some guru meditation.