XMavlink is now derived from the new XObject.  XGLObject is derived directly from XObject.  Thus XGLShapes can now have non-XGLShape objects as children.
A few dynamic_cast<XGLShape *> were needed to prevent the XGL code from trying to render non XGLShape children.

With these changes, MavlinkDevBuildScent can now attach it's XMavlink instance as a child to the shape that is altered by the ATTITUDED message, and at shutdown time the XMavlink instance is now guaranteed to be deleted first, BEFORE the XGLShape is destroyed.  This fixes the intermittent crash on app shutdown.