Using glm to generate a rotation matrix...
After a day of searching for "quaternion to Euler" code,
I found something that suggested glm::quat to glm::mat4.
Tried that and it seems to work.  Now the only thing left
is sensor calibration.  Seems the gyro has drift.  The
literature suggests that each sensor has a "gain vector", ie:
each axis is an analog sensor and therefore will not be ideal
in it's full scale swing.  Meaning each sensor needs to be calibrated.

Not sure I buy that completely, but it's late and I'm too tired to argue