Fix for buggy MadgwickAHRS code...
So. Turns out the open source code for the MadgwickAHRS code
is buggy.  I found this forumn discussion with a fix:

http://diydrones.com/forum/topics/madgwick-imu-ahrs-and-fast-inverse-square-root?id=705844%3ATopic%3A1018435&page=4#comments

which I've added to the code.  However, that's only in the magnetometer code,
IMU only doesn't suffer from the same problem. (supposedly)  I don't much
care about the mag sensor at this point, so maybe it's moot.

It seems to take quite some time to align to the mag data when I use it,
whereas IMU only locks in quickly.

Also, the invsqrt() assumed a 32-bit architecture.  I fixed that per another
comment in the same forum discussion, and suddenly the algorithm behaves as expected.

Setting the "beta" gain to smaller numbers causes increased settling time whilst
eliminating jitter in the resulting quaternion.  Other comments in the same
discussion hint at dynamic modification of the "beta" gain.  Interesting idea.