Madgwick enhancements: dynamic "beta" gain WIP...
First steps toward dynamic "beta" gain.  Idea is that when
the gyro is indicating movement, increase the gain.  Or
when the gyro is NOT indicating movment, decrease the gain.

Added "gryoRateChange" and "accelRateChange", as those are easily
deduced from the existing code. Graphing them in the visualization
to prove they do what I intended.