Enhancement to Madgwick code: adaptive beta gain...
Trial & error on getting adaptive "beta" to work, but it is
doing what I want now.  That is, when gyro detects motion, bump
up the "beta" coefficient to "maxBeta" and allow it to fade gradually
to "betaDef" (the minimum).  This speeds up the settling time
dramatically when rapid changes in orientation occur.