Add "headPosition" to XGLHmd...
... "headPosition" tracks the base offset of the HMD in world space.
That is, it's a virtual camera position that can be manipulated to give
the illusion of moving within the virtual world space.  The hand poses
must have this "headPosition" added to them in order to keep the world objects
that track the hands proximate with the HMD.  Ie: keep your hands near you at all times.