Fixed layerShapes bug...
...turns out I actually implemented layerShapes properly, except for one small detail.
When AddShape() calls CreateShape(), it has to pass it's "layer" argument.  If it doesn't
CreateShape() will use IT'S default "layer" which is not what we want.  Turns out the XGLShapesMap
was already storing pointers to XGLShader objects rather than references, which is the right thing.
Had to tweak XGL destructor just a bit to deal with XGLShader deletion differently.