Refactor animation of XGLShapes...
... used to be that a shapes' animation function would be called right before rendering,
during the RenderScene() phase.  But... attempting to animate the HMD sled revealed a weakness.
Now what happens is the entire shape hierarchy is traversed to apply ALL shape animation functions
(if they exist) before anything is rendered.  So instead of repeat(nextShape,animate it, render it)
we now to repeat(nextShape, animate it) then repeat(nextShape, render it).