Flight controls done right!!!...
Turns out I was REALLY close.  The key was taking the *inverse* of the hmdSled's model
matrix for the HMD view modifier.

So the "rootShape" hack has been reverted, the hmdSled restored, and it's now doing all
that was expected.  AND... lighting of the hand objects is correct.  But the toroidal shapes
are not light correctly. I *think* that's because I need to attach the XGL camera object to the sled.