Getting ready for better XObject smart pointer utilization...
... the 'parent' field of XObject is problematic.  I want to delete
all entries in a std::vector and/or std::map when the vector or map
itself is deleted.  std::smart_ptr (or std::unique_ptr) will make
possible, but the 'parent' doesn't fit with smart pointer semantics.
Turns out only the GUI code uses it, and only in one place.  So 'parent'
has been moved to XGLShape, for now, until I figure out a better way.