XAVStream for audio now uses XCircularBuffer...
...and XAL buffering switched to 512 buffers of 128 samples,
from 64 buffers of 1024 samples.  1024 samples is what the decoder likes
to spit out, but we want to feed the audio decoder (OpenAL) the smallest
chunks we can.  By doing so, we get smoother sense of where audio clock,
which can in turn serve as the decoder sync clock. WIP still, but definite progress.