Got AVPacket situation figured out:...
You CAN build your own.  Just use only the data & size fields of AVPacket.
Copying input packet to "video" packet and now suddenly the CPU usage is
averaging 10% for 1080p60, as low as I've seen in a while.

Also: moved the camera in Mp4DevBuildScene, so the video quad fills the view.