Fixed debug configuration crash in FreeTypeDev...
...GL_CHECK() isn't applied during Release config builds, thus
an erroneous contourOffsets list was built which caused glDrawArrays()
to get a count of -1 went undetected until I switched to debug builds.
The MoveTo() method of XGLFreeType() is modified to detect not only
if we've not seen a vertex yet, but ALSO if the current vertex is
the beginning of a new glyph.  FreeTypeDecomposer() thus handles
a string of text glyphs, instead of just a single glyph.