Add "forceChannels" to AddTexture(std::string)...
... my new Icon files aren't parsed correctly by the SOIL image library.
I don't know why. In diagnosing it I ran across a bug report on github
that mentioned how to use the "forceComponents" argument to SOIL_load_image(),
I tried it and it gave me what I'm looking for. That necessitated changing the
aforementioned AddTexture() call. The new additional arg has a default of zero,
which if passed through is the same as "false".  It's a bit of a smelly hack, but
I don't much care at this point.